requirements.txt
Flask
Flask-Cors
pymongo
python-dotenv
pysmb
